# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/subscriptions/usage": {
      "get": {
        "operationId": "SubscriptionsController_getUsage",
        "summary": "Get remaining requests for today",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Daily limit, requests used, and requests remaining"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/providers": {
      "post": {
        "operationId": "ProvidersController_create",
        "summary": "Add a new AI provider configuration (OpenAI / Claude / Gemini)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProviderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Provider created; apiKeyPreview shows a masked key"
          },
          "400": {
            "description": "Validation failed (missing/short apiKey, invalid enum, etc.)"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "tags": [
          "AI Providers"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "get": {
        "operationId": "ProvidersController_findAll",
        "summary": "List the current user's configured AI providers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Array of provider configs (keys never exposed)"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "tags": [
          "AI Providers"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/providers/{id}": {
      "get": {
        "operationId": "ProvidersController_findOne",
        "summary": "Get a single provider configuration",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider found"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Provider not found or not owned by this user"
          }
        },
        "tags": [
          "AI Providers"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "patch": {
        "operationId": "ProvidersController_update",
        "summary": "Edit a provider configuration (label, model, key rotation, etc.)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProviderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider updated"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Provider not found or not owned by this user"
          }
        },
        "tags": [
          "AI Providers"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "ProvidersController_remove",
        "summary": "Delete a provider configuration",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider removed"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Provider not found or not owned by this user"
          }
        },
        "tags": [
          "AI Providers"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/providers/{id}/enable": {
      "patch": {
        "operationId": "ProvidersController_enable",
        "summary": "Enable a provider",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider enabled"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Provider not found or not owned by this user"
          }
        },
        "tags": [
          "AI Providers"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/providers/{id}/disable": {
      "patch": {
        "operationId": "ProvidersController_disable",
        "summary": "Disable a provider",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider disabled"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Provider not found or not owned by this user"
          }
        },
        "tags": [
          "AI Providers"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/admin/dashboard": {
      "get": {
        "operationId": "AdminController_dashboard",
        "summary": "Aggregate dashboard statistics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Counts of users, subscriptions, and usage"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Authenticated but not an ADMIN"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/stats": {
      "get": {
        "operationId": "AdminController_stats",
        "summary": "Aggregate dashboard statistics (alias)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Counts of users, subscriptions, and usage"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Authenticated but not an ADMIN"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/users": {
      "get": {
        "operationId": "AdminController_users",
        "summary": "List all users (paginated)",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of users with subscription info"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Authenticated but not an ADMIN"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/users/{id}": {
      "get": {
        "operationId": "AdminController_userDetail",
        "summary": "Get a single user (admin view)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Authenticated but not an ADMIN"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/users/{id}/role": {
      "patch": {
        "operationId": "AdminController_setRole",
        "summary": "Change a user's role (USER / ADMIN)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Role updated"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Authenticated but not an ADMIN"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/users/{id}/suspend": {
      "patch": {
        "operationId": "AdminController_suspend",
        "summary": "Suspend (deactivate) a user account",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account suspended"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Authenticated but not an ADMIN"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/users/{id}/reactivate": {
      "patch": {
        "operationId": "AdminController_reactivate",
        "summary": "Reactivate a suspended user account",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account reactivated"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Authenticated but not an ADMIN"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/subscriptions": {
      "get": {
        "operationId": "AdminController_subscriptions",
        "summary": "Subscription overview across all users, grouped by plan and status",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Aggregate counts per plan/status combination"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Authenticated but not an ADMIN"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/subscriptions/{userId}": {
      "get": {
        "operationId": "AdminController_getUserSubscription",
        "summary": "Get a specific user's full subscription record",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription found"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Authenticated but not an ADMIN"
          },
          "404": {
            "description": "User has no subscription record"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "patch": {
        "operationId": "AdminController_overrideSubscription",
        "summary": "Override a specific user's subscription (plan, status, or daily limit)",
        "description": "Direct admin control distinct from the self-service /subscriptions/* endpoints - e.g. to comp a Premium plan, correct a failed payment webhook, or force-cancel for abuse.",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OverrideSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subscription updated"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Authenticated but not an ADMIN"
          },
          "404": {
            "description": "User has no subscription record"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/providers": {
      "get": {
        "operationId": "AdminController_providers",
        "summary": "AI provider overview across all users, grouped by provider and status",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Aggregate counts per provider/enabled combination"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Authenticated but not an ADMIN"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/users/{userId}/providers": {
      "get": {
        "operationId": "AdminController_listUserProviders",
        "summary": "List every AI provider a specific user has configured",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of the user's providers (keys never exposed)"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Authenticated but not an ADMIN"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/providers/{id}/disable": {
      "patch": {
        "operationId": "AdminController_disableProvider",
        "summary": "Force-disable any user's provider (e.g. reported abuse, leaked key)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider disabled"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Authenticated but not an ADMIN"
          },
          "404": {
            "description": "Provider not found"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/providers/{id}/enable": {
      "patch": {
        "operationId": "AdminController_enableProvider",
        "summary": "Re-enable any user's provider",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider enabled"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Authenticated but not an ADMIN"
          },
          "404": {
            "description": "Provider not found"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/providers/{id}": {
      "delete": {
        "operationId": "AdminController_deleteProvider",
        "summary": "Force-delete any user's provider configuration",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider removed"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Authenticated but not an ADMIN"
          },
          "404": {
            "description": "Provider not found"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/usage-analytics": {
      "get": {
        "operationId": "AdminController_usageAnalytics",
        "summary": "API usage analytics grouped by day",
        "parameters": [
          {
            "name": "days",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Per-day request counts, error counts, avg duration"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Authenticated but not an ADMIN"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/usage": {
      "get": {
        "operationId": "AdminController_usage",
        "summary": "API usage analytics grouped by day (alias)",
        "parameters": [
          {
            "name": "days",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Per-day request counts, error counts, avg duration"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Authenticated but not an ADMIN"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/logs": {
      "get": {
        "operationId": "AdminController_logs",
        "summary": "Paginated raw request logs",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated ApiUsageLog rows"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Authenticated but not an ADMIN"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "EchoGPT Backend API",
    "description": "Production-ready backend for the EchoGPT Chrome Extension: auth, subscriptions, AI provider management, chat, web search, and admin analytics.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateProviderDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "enum": [
              "OPENAI",
              "CLAUDE",
              "GEMINI"
            ]
          },
          "label": {
            "type": "string",
            "description": "Friendly label, e.g. \"Work OpenAI key\""
          },
          "apiKey": {
            "type": "string",
            "description": "Plaintext API key - encrypted at rest, never returned again"
          },
          "model": {
            "type": "string"
          },
          "baseUrl": {
            "type": "string",
            "description": "Override base URL (self-hosted / proxy endpoints)"
          },
          "isDefault": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "name",
          "apiKey"
        ]
      },
      "UpdateProviderDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string"
          },
          "apiKey": {
            "type": "string",
            "description": "Provide only to rotate the stored key"
          },
          "model": {
            "type": "string"
          },
          "baseUrl": {
            "type": "string"
          },
          "isEnabled": {
            "type": "boolean"
          },
          "isDefault": {
            "type": "boolean"
          }
        }
      },
...
      "UpdateUserRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN"
            ]
          }
        },
        "required": [
          "role"
        ]
      },
      "OverrideSubscriptionDto": {
        "type": "object",
        "properties": {
          "plan": {
            "type": "string",
            "enum": [
              "FREE",
              "PREMIUM"
            ],
            "description": "Force the user's plan"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "CANCELED",
              "EXPIRED",
              "PAST_DUE"
            ],
            "description": "Force the user's subscription status"
          },
          "dailyLimit": {
            "type": "number",
            "description": "Override the daily request limit",
            "minimum": 0
          }
        }
      }
    }
  }
}